Implement functions, lambdas. Improve traces.
Adds context to each stack trace item when evaluating expressions, including a little bit of "why" for what's going on, which gives pretty good info in the stack trace even when Ruamel refuses to include code snippets.

Adds support for custom Python functions and lambdas within Yamlet, and adds an example of each to the main `example.py` file.