Tidy up, add `explain_value`, update README.
Adds provenance info for scalars on top of compositing history. This is worth documenting as an improvement over straight BCL.

Tidies up and does a little refactoring; I should still create a wrapper over Ruamel marks at some point, as using str() on them and manipulating the result is a bit ugly. For now, though, this does what it needs to do, so it's time to put a bow on it.

Talks the API up a bit more in the readme, too.