Hammer it some more.
I had to implement pretty much everything from GCL to make this work. I'm gonna have to finish generalizing the compositable API in order to drag this to the finish line.

Then I need some tests for all the new goodies I just bolted to the system. Plus I'd better make null work the way it's supposed to.

This... was a tall order I placed.