Finishes preprocessors.
Adds tests for the newly added constructs (external and null).

Another weird gotcha is that even constant expressions (such as `!expr null`) are left as deferred values, which interfers with the way they are composited. It would probably be better to add an `is_constant()` routine (similar to the `is_null()` routine to pre-evaluate as much as possible. Anything with name refs or func calls is not constant.

Another missing feature from GCL is local values, but I think this is a fine place to draw the line for right now.