Fix `!else:`, fix `load` vs `loads`, add main example to unit tests.
I had meant `load()` to take a filename, while `loads()` takes a string or stream, but that's not how APIs that differentiate between `load` and `loads` work. So now `loads` is `load` and continues to accept string or stream, while the original `load` has been renamed to `load_file` and continues to accept a filename.

I added a nasty-ish hack to preprocess the Yaml sources so that any lone `!else:` on a line would be replaced with `!else :` so that parse would succeed. This will break any values that naturally contain a line with nothing but `!else:` (and maybe a yaml-looking comment) on them. This shouldn't really happen in practice, but it bothers me that it's an issue. I will reach out to the YAML and Python communities about next steps.

Also adds the main example file to the unit test suite, since I've managed to break it a couple times now.