Two new features: Warnings on undefined, module_vars.
Yamlet now supports printing warnings (in GitHub Actions format) when a variable is undefined and its default value is used. This is in addition to the earlier support for exception message prefixes to allow formatting built-in exceptions as errors. It would probably behoove me to just add a warning/error formatter function, but we're in the moving fast phase.

Also adds `module_vars`, which can be returned as part of a new `ImportInfo` class in place of a filename Path (or string) from the import resolver, which allows dynamically setting globals on a per-import basis.